backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
numpy>=1.21.0
opencv-python>=4.5.0
mediapipe>=0.10.0
tensorflow>=2.13.0
pandas>=1.3.0
matplotlib>=3.4.0
scikit-learn>=0.24.0
Pillow>=8.0.0
python-dotenv>=0.19.0
